// src/api/generated/model/getTransactions200DataWalletTransactionsItem.ts
/**
 * Generated by orval v7.16.1 🍺
 * Do not edit manually.
 * Smart City Hub
 * A comprehensive API
 * OpenAPI spec version: 1.0.0
 */

export type GetTransactions200DataWalletTransactionsItem = {
  id: number;
  /** @nullable */
  wallet_id: number | null;
  /** @nullable */
  transaction_type: string | null;
  amount: number;
  /** @nullable */
  target_wallet_id: number | null;
  /** @nullable */
  target_service: string | null;
  /** @nullable */
  description: string | null;
  created_at: string;
};