// src/api/generated/model/getTransactions200DataCardTransactionsItem.ts
/**
 * Generated by orval v7.16.1 🍺
 * Do not edit manually.
 * Smart City Hub
 * A comprehensive API
 * OpenAPI spec version: 1.0.0
 */

export type GetTransactions200DataCardTransactionsItem = {
  id: number;
  card_id: number;
  /** @nullable */
  card_type: string | null;
  /** @nullable */
  transaction_type: string | null;
  /** @nullable */
  transaction_category: string | null;
  /** @nullable */
  reference: string | null;
  amount: number;
  /** @nullable */
  description: string | null;
  created_at: string;
};