// src/api/generated/model/getMetroCardsUserUserId200DataMetroCardsItem.ts
/**
 * Generated by orval v7.16.1 🍺
 * Do not edit manually.
 * Smart City Hub
 * A comprehensive API
 * OpenAPI spec version: 1.0.0
 */
import type { GetMetroCardsUserUserId200DataMetroCardsItemStatus } from './getMetroCardsUserUserId200DataMetroCardsItemStatus';

export type GetMetroCardsUserUserId200DataMetroCardsItem = {
  id: number;
  user_id: number;
  balance?: number;
  card_number: string;
  status: GetMetroCardsUserUserId200DataMetroCardsItemStatus;
  created_at: string;
  updated_at: string;
};